fixed patch function
refetch working
